reworked fps component and update loop
fps component was too complex, had to make it simpler that it basically does what it it's supposed to do, calculate the FPS and let the rendering/showing text to their respective components.

added a variable for the components to have an owner so components such as fps component can make use of the text component.